# Hash now uses an open addressing algorithm