@lribeiro 2019-07-31
This gives a nice speedup on our internal app for log processing!

@bcardiff will this be considered for 0.30?

Time in seconds

|test|0.29|0.30dev|speedup|
|---|---:| ---:| ---:|
|forti|7.353|6.722|9.4%|
|forti_kv|5.462|4.867|12,2%|
|dns_grt|4.714|4.466|5,6%|


Details 

`~/Work/crystal-1/bin/crystal version`
```bash
Using compiled compiler at `.build/crystal'
Crystal 0.30.0-dev [ec2a26f9a] (2019-07-31)

LLVM: 8.0.0
Default target: x86_64-apple-macosx
```
```
~/Work/crystal-1/bin/crystal build --release -o bin/log_clean src/log_clean.cr 
Using compiled compiler at `.build/crystal'
bin/log_clean -c test/forti/forti.yml >    /dev/null  6.83s user 1.17s system 119% cpu 6.722 total
bin/log_clean -c test/forti/forti_kv.yml > /dev/null  5.11s user 2.02s system 146% cpu 4.867 total
bin/log_clean -c config/dns_grt.yml >      /dev/null  4.50s user 0.74s system 117% cpu 4.466 total

```

`crystal version`
```bash
Crystal 0.29.0 (2019-06-06)

LLVM: 6.0.1
Default target: x86_64-apple-macosx
```
```
bin/log_clean -c test/forti/forti.yml >    /dev/null  7.49s user 1.26s system 119% cpu 7.353 total
bin/log_clean -c test/forti/forti_kv.yml > /dev/null  5.80s user 2.08s system 144% cpu 5.462 total
bin/log_clean -c config/dns_grt.yml >      /dev/null  4.78s user 0.79s system 118% cpu 4.714 total
```