@yxhuvud 2019-07-31
I wish github had properly threaded review comments. This is a comment to my own review: I note that you get lets of repeated allocation warnings and that one CI fail due to out of memory.