@yxhuvud 2019-07-31
@straight-shoota Are all the warnings about large allocations also old?