@funny-falcon 2019-07-31
Rather clean code! I liked to read it.