@asterite 2019-07-31
@funny-falcon 

> Rather clean code! I liked to read it.

Thank you! 😊 And thank you for the thorough review, you found many good things to fix.