@funny-falcon 2019-07-31
`Hash.shift` is rather useful for:
- simple LRU implementation
- keyed FIFO queue
Given it is quite easy to add @first , it is better to add it than to not.

> 31 июля 2019 г., в 21:35, Ary Borenszweig <notifications@github.com> написал(а):
> 
> @asterite commented on this pull request.
> 
> In src/hash.cr:
> 
> >      end
>    end
>  
>    private module BaseIterator
> -    def initialize(@hash, @current)
> +    def initialize(@hash)
> +      @index = 0
> Ah, yes, I see what you mean. I think we should either remove Hash#shift or implement your suggestion. We copied Hash#shift from Ruby but I don't know what's a real use for that it.
> 
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub, or mute the thread.