@rumenzu 2019-07-31
A quick [benchmark](https://github.com/rumenzu/quartz/blob/master/benchmarks/devstone.cr) from my simulation framework reveals a really nice speedup of **~1.37x**!
Hash was already identified as a bottleneck in my case after profiling.

`./devstone 10000 3` results:

Crystal version | Wall clock time (s)
--------------- | --------------------
Crystal 0.30.0-dev [008647480] (2019-07-31) LLVM: 8.0.0 | 7.797805118
Crystal 0.29.0 (2019-06-06) LLVM: 6.0.1 | 5.703113211