@asterite 2019-07-31
> Given it is quite easy to add @FIRST , it is better to add it than to not.

Yes, but it adds 4 bytes for all hashes even if you never use `shift`. Same reason why we didn't do that for `Array`. If you need `shift` like that you can use `Deque`, which is in the standard library and it's more efficient (at the cost of a bit more memory).

Said another way, we can consider this as a separate issue.