@yxhuvud 2019-07-31
@asterite: I think the warnings isn't due to requiring more memory but because libgc doesn't like big allocations. That is, it is perfectly happy to allocate a GB if you do it in 4KB chunks, but if you do it all at once you get a warning. So doing fewer tests wouldn't really help with the warnings as you would still get one once the amount of entries in a hash get big enough.