@asterite 2019-07-31
@yxhuvud Makes sense. I think I added a couple of tests that exercise some big-ish hashes.