@asterite 2019-07-31
@funny-falcon On second thought it seems we can put the 4 bytes of `@first` in a place where it doesn't increase the size of `Hash`, at least in 64 bits where everything is aligned to 8 bytes. And the same is true for Array! So I think I'll add those (I'll add `@first` for Hash in this PR, and for Array in a separate PR). It'll be cool to have O(1) `shift`. I'm thinking for Array I'll just increment `@first` but also clear the value behind that pointer. Though I'll probably wait until parallelism or coordinate with @bcardiff because I know he's been doing some changes to Array to make it a bit more safe in parallel environments.