@asterite 2019-07-31
Now with optimized `shift`, thanks to @funny-falcon 's suggestion.

Benchmark. Code:

```crystal
old_hash = OldHash(Int32, Int32).new
new_hash = Hash(Int32, Int32).new

p instance_sizeof(typeof(old_hash))
p instance_sizeof(typeof(new_hash))

100_000.times do |i|
  old_hash[i] = i
  new_hash[i] = i
end

time = Time.now
while old_hash.shift?
end
puts "Old: #{Time.now - time}"

time = Time.now
while new_hash.shift?
end
puts "New: #{Time.now - time}"
```

Results:

```
56
56
Old: 00:00:02.272458000
New: 00:00:00.000244000
```

(meaning the bytesize of a Hash didn't change with this optimization)