@RX14 2019-07-31
This is absolutely amazing! I think robin hood hashing would be a neat future optimization