@funny-falcon 2019-08-01
It is better to maintain `@first` in `delete_entry_and_update_counts`. It is single point of all deletions.
And it is better not simply `@first+=1` but:

```ruby
delete_entry(index)
@size-=1
@deleted_count+=1
return if index != @first
if @size == 0
  @first = @deleted_count
else
  @first += 1
  @first += 1 while @entries[@first].deleted?
end
```

And then first_entry could be optimized as well:
```ruby
@size == 0 ? nil : @entries[@first]
```

And shift

```ruby
if first_entry
  delete_entry_and_update_counts(@first)
  {first_entry.key, first_entry.value}
else
```