@ysbaddaden 2019-08-01
You may also consider the [Robin Hound Hashing](https://cs.uwaterloo.ca/research/tr/1986/CS-86-14.pdf) (site may not be responding) optimization. Basically: 
> when you probe for a position to insert a new element, if the probe length for the existing element is less than the current probe length for the element being inserted, swap the two elements and keep going.

- [Robin Hood Hashing should be your default Hash Table implementation](https://www.sebastiansylvan.com/post/robin-hood-hashing-should-be-your-default-hash-table-implementation/)

Plus the *Learn Hash Table the Hard Way* series that talks about probes (linear, quadratic, double hash) and Robin Hound Hashing:
- [Part 1: Probe Distributions](http://www.idryman.org/blog/2017/07/04/learn-hash-table-the-hard-way/)
- [Part 2: Probe Distributions With Deletions](http://www.idryman.org/blog/2017/07/18/learn-hash-table-the-hard-way-2/)
- [Part 3: Probe Distributions and Run Time Performance](http://www.idryman.org/blog/2017/08/06/learn-hash-table-the-hard-way-3/)